chore: install command

+++ packages/core/server/src/commands/install.ts
@@ -10,34 +10,30 @@ export default (app: Application) => {
     .option('-r, --retry [retry]')
     .option('-I, --ignore-installed')
     .action(async (...cliArgs) => {
-      let installed = false;
+      const installed = await app.isInstalled();
+
       const [opts] = cliArgs;
 
-      if (opts.ignoreInstalled) {
-        if (await app.isInstalled()) {
-          console.log('Application installed');
-          return;
-        }
+      if (opts.ignoreInstalled && installed) {
+        app.logger.info('NocoBase is already installed. Ignore install command.');
+        return;
       }
 
-      if (!opts?.clean && !opts?.force) {
-        if (await app.isInstalled()) {
-          installed = true;
-          if (!opts.silent) {
-            console.log('NocoBase is already installed. To reinstall, please execute:');
-            console.log();
-            const command = '$ yarn nocobase install -f';
-            console.log(chalk.yellow(command));
-            console.log();
-            console.log(chalk.red('This operation will clear the database!!!'));
-            console.log();
-          }
-          return;
+      if (!opts?.clean && !opts?.force && installed) {
+        if (!opts.silent) {
+          console.log('NocoBase is already installed. To reinstall, please execute:');
+          console.log();
+          const command = '$ yarn nocobase install -f';
+          console.log(chalk.yellow(command));
+          console.log();
+          console.log(chalk.red('This operation will clear the database!!!'));
+          console.log();
         }
+        return;
       }
 
       if (!opts.silent || !installed) {
-        console.log(`Start installing NocoBase`);
+        app.logger.info(`Start installing NocoBase`);
       }
 
       await app.install({
